+++ R/tune_args.R
@@ -113,7 +113,8 @@ tune_id <- function(x) {
 
     # [tune()] will always return a call object
     if (is.call(x)) {
-      if (rlang::call_name(x) == "tune") {
+      nm <- rlang::call_name(x)
+      if (!is.null(nm) && nm == "tune") {
         # If an id was specified:
         if (length(x) > 1) {
           return(x[[2]])